working version ready for vogsphere

+++ srcs/sp_cases.c
@@ -6,7 +6,7 @@
 /*   By: jheiskan <jheiskan@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/14 18:49:09 by jheiskan          #+#    #+#             */
-/*   Updated: 2022/03/07 13:56:25 by jheiskan         ###   ########.fr       */
+/*   Updated: 2022/04/20 16:35:14 by jheiskan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ ssize_t	special_cases_di(t_printf *tab)
 		return ((long int)va_arg(tab->args, ssize_t));
 	else if (tab->h)
 		return ((short int)va_arg(tab->args, ssize_t));
-	return (va_arg(tab->args, int));
+	return ((int)va_arg(tab->args, ssize_t));
 }
 
 long double	special_cases_f(t_printf *tab)
